+++ server/api/announcement/announcement.controller.js
@@ -1,0 +1,6 @@
+'use strict';
+
+// Gets a list of Announcements
+export function index(req, res) {
+  res.json([]);
+}