+++ server/config/seed.js
@@ -4,7 +4,7 @@
  */
 'use strict';
 import config from './environment/';
-import { User } from '../sqldb';
+import { User, Announcement } from '../sqldb';
 import Sequelize from 'sequelize';
 
 export default function seedDatabaseIfNeeded() {
@@ -111,7 +111,28 @@ export default function seedDatabaseIfNeeded() {
             }
           }
         ])
-      .then(() => console.log('Seeded users')))])
+      .then(() => console.log('Seeded users'))),
+
+      Announcement.destroy({
+        where: {
+          $and: [
+            { section: {$eq: 'Sunday, April 16th Class Schedule'} },
+            { title: {$eq: 'East Libery School'} }
+          ]
+        }
+      })
+      .then(() =>
+        Announcement.bulkCreate([
+          {
+            section: 'Sunday, April 16th Class Schedule',
+            title: 'East Libery School',
+            description: '- all classes running as scheduled',
+            expires: '2017-04-30T00:00:00.000-05:00'
+          }
+        ])
+      )
+      .then(() => console.log('Seeded announcements'))
+    ])
     .then(() => console.log('Seeded database.'))
     .catch(err => console.log('Seeding error: ', err));
   }