+++ server/sqldb/index.js
@@ -15,5 +15,6 @@ let db = {
 // Insert models below
 db.User = db.sequelize.import('../api/user/user.model');
 db.Order = db.sequelize.import('../api/order/order.model');
+db.Announcement = db.sequelize.import('../api/announcement/announcement.model');
 
 module.exports = db;